Add files via upload
Week 3 Homework assignment. Wanted to add opacity changes to the background image, but wasn't able to successfully execute without corrupting the layout or the audio alert. Additionally, explanatory notations in some of the code delayed the audio until after the user clicked 'OK' on the alert box. I will post an updates version as soon as I find the solution to the problem.